Implement signin() and getNextClass()
signin calls a google auth api that returns an accessToken when a user gives permission to the app. This token is stored in state. GetNextClass makes one api call to find the school calendar and makes another api call to find the next class in the calendar. There is conditional rendering as well depending if the user is logged in or not.